LayoutTests/imported/w3c/web-platform-tests/css/cssom/cssstyledeclaration-mutationrecord-001-expected.txt


PASS CSSStyleDeclaration.setPropertyValue queues a mutation record when serialization is changed
